Deleted OAuth2UserDetailsAdapater and made it an anonymous class instead. Works fine with oaut2 and non oauth2
