--- sapl-server-ce/src/main/java/io/sapl/server/ce/security/oauth2/OAuth2UserDetailsAdapter.java
@@ -1,65 +1,0 @@
-/*
- * Copyright (C) 2017-2024 Dominic Heutelbeck (dominic@heutelbeck.com)
- *
- * SPDX-License-Identifier: Apache-2.0
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.sapl.server.ce.security.oauth2;
-
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-
-import lombok.experimental.Delegate;
-
-public class OAuth2UserDetailsAdapter implements OAuth2User, UserDetails {
-
-    private static final long serialVersionUID = 4075942975439541721L;
-
-    @Delegate(types = OAuth2User.class)
-    private transient OAuth2User oauth2User;
-
-    public OAuth2UserDetailsAdapter(OAuth2User oauth2User) {
-        this.oauth2User = oauth2User;
-    }
-
-    @Override
-    public String getPassword() {
-        return "PASSWORD NOT NEEDED";
-    }
-
-    @Override
-    public String getUsername() {
-        return getName();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return oauth2User.getAuthorities() != null;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
